chore: rename `contrib` to `lib`
