+++ packages/core/src/engine/BBox.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "vitest";
-import { compDict } from "../contrib/Functions.js";
-import { numsOf } from "../contrib/Utils.js";
+import { compDict } from "../lib/Functions.js";
+import { numsOf } from "../lib/Utils.js";
 import { makeCircle } from "../shapes/Circle.js";
 import { makeEllipse } from "../shapes/Ellipse.js";
 import { makeImage } from "../shapes/Image.js";
