+++ packages/core/src/lib/Curves.ts
