+++ packages/core/src/lib/ImplicitShapes.ts
