+++ packages/core/src/lib/Queries.ts
