+++ packages/core/src/lib/__tests__/ClosestPoint.test.ts
