+++ packages/core/src/lib/__tests__/CurveConstraints.test.ts
