+++ packages/core/src/lib/__tests__/Queries.test.ts
